+++ analysis/studyTasks/01_buildCohorts.R
@@ -52,7 +52,7 @@ analysisSettings <- readSettingsFile(here::here("analysis/settings/strata.yml"))
 
 # E. Script --------------------
 
-startSnowflakeSession(con = con, executionSettings = executionSettings)
+#startSnowflakeSession(con = con, executionSettings = executionSettings)
 
 ## Initialize cohort tables
 initializeCohortTables(executionSettings = executionSettings, con = con, dropTables = TRUE)